pyproject.toml
[project]
name = "tooling-postgres-cdc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=37.6.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.12.12",
    "ty>=0.0.1a20",
]
